requirements.txt
flask
flask_sqlalchemy
python-dotenv
faker

# Optional but highly recommended:
flask-wtf          # For form handling and CSRF protection
wtforms            # If using standalone WTForms
email-validator    # If validating email fields in forms
pytest             # For unit testing
pytest-flask       # Flask-specific pytest helpers
